feat: add solution to problem cjsToEsm.mjs
